TabIndex order change, focus styles, styling page
Map of the detail page is set under the reviews and other elements.
So the focus won't go first to the map.
The elements on the main page are centered.
Footer of main page is styled like the detail page

Images are optimized and alt tag is added to them.
Images will scale better without stretch.