Service worker implemented
New file called sw.js is added.
In here you can find the code where fetches are intercepted and stored
in the cache.
If there is a cache of the file, the cache will be retured instead of a
server.
Javascript and css that are loaded, will also be stored in the cache so
when the page reloads these files won't be loaded from the server.

Further minor changes:
-The viewport is added to the head of the html for scalability on mobile
devices.