Refactoring
The favourite state will also be stored when the user is offline.

The serverice worker has been refactored, useless code has been removed
and made it more performant.

Instead of xhttp use fetch to make API calls.

Insert all of the request code in the dbhelper.

The API request will be stored in the local storage when user is offline.
If a request is being made the code will loop through the local storage
and will try to make the requests again.
The user will always be able to resend all of the requests bu pressing
the button on the status bar.

Feature

Adding a favicon icon tho the site.
Gulp will now also compile all of the icons.